Write initial changes to be compatible with WindTunnel project.
