+++ anemoi/__init__.py
@@ -1,5 +1,7 @@
+
 from .analytical import AnalyticalHelmholtz
-from .discretization import BaseDiscretization
+from .discretization import BaseDiscretization, DiscretizationWrapper
 from .minizephyr import MiniZephyr, MiniZephyr25D
 from .eurus import Eurus
 from .source import BaseSource, SimpleSource, StackedSimpleSource, KaiserSource, FakeSource
+from .meta import BaseModelDependent, AttributeMapper
