+++ anemoi/discretization.py
@@ -1,7 +1,8 @@
 
-from .meta import BaseModelDependent
-from .solver import DirectSolver
+import copy
 import numpy as np
+from .meta import AttributeMapper, BaseModelDependent
+from .solver import DirectSolver
 
 class BaseDiscretization(BaseModelDependent):
     
@@ -42,3 +43,47 @@ class BaseDiscretization(BaseModelDependent):
     
     def __call__(self, value):
         return self*value
+
+class DiscretizationWrapper(AttributeMapper):
+    
+    initMap = {
+    #   Argument        Required    Rename as ...   Store as type
+        'disc':         (True,      None,           None),
+        'scaleTerm':    (False,     '_scaleTerm',   np.complex128),
+    }
+    
+    maskKeys = []
+    
+    def __init__(self, systemConfig):
+        
+        super(DiscretizationWrapper, self).__init__(systemConfig)
+        self.systemConfig = {key: systemConfig[key] for key in systemConfig if key not in self.maskKeys}
+    
+    @property
+    def scaleTerm(self):
+        return getattr(self, '_scaleTerm', 1.)
+    
+    @property
+    def _spConfigs(self):
+        
+        def duplicateUpdate(spu):
+            nsc = copy.copy(self.systemConfig)
+            nsc.update(spu)
+            return nsc
+        
+        return (duplicateUpdate(spu) for spu in self.spUpdates)
+    
+    @property
+    def subProblems(self):
+        if getattr(self, '_subProblems', None) is None:
+            
+            self._subProblems = map(self.disc, self._spConfigs)
+        return self._subProblems
+    
+    @property
+    def spUpdates(self):
+        raise NotImplementedError
+    
+    def __mul__(self, rhs):
+        raise NotImplementedError
+
